Core/Creatures: fixed school immunity unapply when updating entry and unified code
- Also use correct data type for SQL (INT <-> uint32)

(cherry picked from commit 00c0ec3f2784e67f4ccc7616bf65aa7e5496687a)

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -540,7 +540,7 @@ void ObjectMgr::LoadCreatureTemplate(Field* fields)
     creatureTemplate.WidgetSetUnitConditionID = fields[72].GetInt32();
     creatureTemplate.RegenHealth            = fields[73].GetBool();
     creatureTemplate.MechanicImmuneMask     = fields[74].GetUInt32();
-    creatureTemplate.SpellSchoolImmuneMask  = fields[75].GetUInt8();
+    creatureTemplate.SpellSchoolImmuneMask  = fields[75].GetUInt32();
     creatureTemplate.flags_extra            = fields[76].GetUInt32();
     creatureTemplate.ScriptID               = GetScriptId(fields[77].GetString());
 }
